feat(sugar): restructure discord webhooks & update typescript to fix issues
