+++ sugar/package.json
@@ -8,13 +8,13 @@
   "license": "MIT",
   "devDependencies": {
     "@types/express": "^4.17.15",
-    "typescript": "^4.9.4"
+    "typescript": "^5.2.2"
   },
   "dependencies": {
     "@beemobot/common": "^1.0.0",
     "@sentry/node": "^7.28.1",
     "chargebee-typescript": "^2.15.0",
-    "discord.js": "^14.7.1",
+    "discord.js": "^14.13.0",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "express-basic-auth": "^1.2.1",
