+++ sugar/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "ES2022",
-    "module": "ESNext",
+    "module": "Node16",
     "moduleResolution": "Node16",
     "allowSyntheticDefaultImports": true,
     "declaration": true,
